moved function to hwmgrclient
