+++ adaptors/dell-hwmgr/hwmgrclient/client.go
@@ -336,3 +336,50 @@ func (c *HardwareManagerClient) GetSecret(ctx context.Context, secretKey string)
 
 	return response.JSON200, nil
 }
+
+// ValidateNodepoolwithResourceSelector validates the hardware manager resource group data with nodepool
+func (c *HardwareManagerClient) ValidateNodepoolWithResourceSelector(
+	ctx context.Context,
+	nodepool *hwmgmtv1alpha1.NodePool,
+	resourceSelector map[string]hwmgrapi.RhprotoResourceSelectorGetResponse,
+) error {
+
+	for _, node := range nodepool.Spec.NodeGroup {
+		nodeName := node.NodePoolData.Name
+
+		if resource, exists := resourceSelector[nodeName]; exists {
+			if node.NodePoolData.HwProfile != *resource.ResourceProfileId {
+				return fmt.Errorf("invalid resource profile for node %s\n expected: %s found: %s",
+					nodeName, node.NodePoolData.HwProfile, *resource.ResourceProfileId)
+			}
+			if float32(node.Size) != *resource.NumResources {
+				return fmt.Errorf("invalid num of resources for node %s\n expected: %f found: %f",
+					nodeName, float32(node.Size), *resource.NumResources)
+			}
+			if node.NodePoolData.ResourcePoolId != *resource.RpId {
+				return fmt.Errorf("invalid resource pool id for node %s\n expected: %s found: %s",
+					nodeName, node.NodePoolData.ResourcePoolId, *resource.RpId)
+			}
+			// Find value for key role in included labels
+			role := ""
+			if resource.Filters != nil && resource.Filters.Include != nil {
+				for _, label := range *resource.Filters.Include.Labels {
+					if label.Key != nil && *label.Key == "role" {
+						role = *label.Value
+						break
+					}
+				}
+			}
+			// Comparing role to node name
+			if node.NodePoolData.Name != role {
+				return fmt.Errorf("invalid role for node %s\n expected: %s found %s",
+					nodeName, node.NodePoolData.Name, role)
+			}
+
+		} else {
+			return fmt.Errorf("validation failed, %s node does not exist in resource group", nodeName)
+		}
+
+	}
+	return nil
+}
